Removed previous file push v1 files, and renamed file push v2 to be without the v2 suffix.
